Move OHLC items into new module
